move VSS and disk discovery back into ntfs accessor
